move `utils` module
